refactor: change folder to navigationBar and repath
