+++ src/app/(root)/setting/page.tsx
@@ -1,7 +1,7 @@
 import { auth } from '@/lib/auth';
 import UpdateInformationForm from '@/components/setting-page/UpdateInformationForm';
 import { redirect } from 'next/navigation';
-import Divider from '@/components/layout/Divider';
+import Divider from '@/components/navigationBar/Divider';
 import VerificationStatus from '@/components/setting-page/VerificationStatus';
 import DeleteAccountButton from '@/components/setting-page/DeleteAccountButton';
 import { AuthProvider } from '@/components/auth/authProvider';
